fix: Update Docker port, add tests AWS credentials and rename exception
Changed the port number in docker-compose.yml from 8000 to 8001 to avoid port conflict. AWS credentials were added in the local test cases to ensure proper authentication during testing. The name for the exception 'NotFound' was updated to 'CoralNotFound' to enhance clear identification and minimize mix up with other NotFound exceptions.