feat: Add TableClass attribute to Coraline and update Python version
Added a new attribute "TableClass" to Coraline to allow specification of the type of the table when creating it. This improves the flexibility of table creation, allowing users to select the desired TableClass depending on their usage pattern. The Python version constraint was also refined from ">=3.9,<=3.11" to ">3.9,<4" for better compatibility and to limit usage to supported Python versions. Furthermore, the Python version utilised was upgraded from "3.10" to "3.11.4" for improved performance and newer features.